Changes to make Ren-C Rebol variation work
This isolates out changes to make the Ren-C version run.

The role previously played by NONE! is now split into two separate
cases.  Signifying absence of any value is done with NULL...which is
a state that cannot be put in blocks, and hence does a better job of
meaning "no, really, nothing".  The role of placeholder is taken by
BLANK!, which has a visual representation of underscore (`_`).

AJOIN is replaced with the more generalized specialization of DELIMIT
called UNSPACED.  The usage of REJOIN is also removed.

The CHAR! datatype no longer exists, as representing characters has
been folded in with "ISSUE!".  This unified type is slated to be called
TOKEN!, as an immutable/indivisible fragment of string material with
optimized representation.