Brute force function handlers into operation definitions
There was some associated Rebol code for the operations that would run
when the operations were matched, in order to implement their
whitespace behavior.  The written-out-by-hand parse rules would make
the connection between the named operation defintition and the named
function by building an instruction block.

(The goal of building the instruction block was to have a capture of
the instruction so it could be dumped...but then it was also an
executable item.)

This tries to glue the function definitions onto the operations, without
worrying too much if doing so is messy.  It does wind up being a rather
messy problem for a few reasons, described in the comments.  These
problems should be studied and remedied.

Rough though this is, it moves closer to the desired goal.