Make parameterization of instructions consistent
Part of the PARSE-driven interpreter concept is that arguments to
the opcodes expressed in the instruction stream get captured as part
of the parse rule to build the instruction.

The process was capturing a parameter in a variable called `param`.
There was then per-opcode handling to build an instruction as a block
which could be dumped, as well as executed to run the associated
implementation code for the instruction.

However, not all instructions were properly composing the (param)
into the instruction.  So some of the implementation functions were
just referring to the temporary variable.  This meant that the
debug purpose of building the instruction wasn't fully served, and
it also meant the instruction interface wasn't consistent with the
function arguments it took.

This corrects the paraameterization so all instructions have the
label or number they intend.