Update for Recent Ren-C Changes
PARSE no longer returns LOGIC!, in accordance with the new idea of
synthesized rules.  Use ELSE instead of testing result directly.

The elimination of /ONLY means that evaluative values must be quoted
if they are to be appended as-is, or put in BLOCK! if literal.  This
applies to derivatives of APPEND like KEEP, as well as functions
like FIND and INSERT.